+++ packages/core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Fix handling of functions with struct or mapping storage pointer arguments.
+
 ## 1.7.3 (2021-05-13)
 
 - Simplify clean up of manifest data before storing to disk.